+++ examples/load-obj-file.js
@@ -1,15 +1,11 @@
 import MatrixEngineWGPU from "../src/world.js";
 import {downloadMeshes} from '../src/engine/loader-obj.js';
-import {geoTypesForMorph, LOG_MATRIX} from "../src/engine/utils.js";
-import {MeshMorpher} from "../src/engine/procedural-mesh.js";
 import {addRaycastsAABBListener} from "../src/engine/raycast.js";
-import {uploadGLBModel} from "../src/engine/loaders/webgpu-gltf.js";
 
 export var loadObjFile = function() {
   let loadObjFile = new MatrixEngineWGPU({
     canvasSize: 'fullscreen',
-    fastRender: 0.4,                // < 1 // recommended
-    // fastRenderAlternative: true, // not good for mobile
+    fastRender: 0.6,
     dontUsePhysics: true,
     mainCameraParams: {
       type: 'WASD',
@@ -43,7 +39,7 @@ export var loadObjFile = function() {
 
     async function onLoadObj(m) {
       loadObjFile.addMeshObj({
-        material: {type: 'mirror'},
+        material: {type: 'standard'},
         position: {x: 0, y: -1, z: -20},
         rotation: {x: 0, y: 0, z: 0},
         scale: [100, 100, 100],
@@ -70,7 +66,7 @@ export var loadObjFile = function() {
       });
 
       loadObjFile.addMeshObj({
-        material: {type: 'mirror'},
+        material: {type: 'minia'},
         position: {x: 0, y: 3, z: -10},
         rotation: {x: 0, y: 0, z: 0},
         rotationSpeed: {x: 0, y: 0, z: 0},
@@ -78,7 +74,7 @@ export var loadObjFile = function() {
         name: 'cube',
         mesh: m.cube,
         envMapParams: {
-          baseColorMix: 0.8,                // CLEAR SKY
+          baseColorMix: 0.99,                // CLEAR SKY
           mirrorTint: [0.9, 0.95, 1.0],     // Slight cool tint
           reflectivity: 0.25,               // 25% reflection blend
           illuminateColor: [0.3, 0.7, 1.0], // Soft cyan
@@ -127,13 +123,13 @@ export var loadObjFile = function() {
 
       loadObjFile.lightContainer[0].intensity = 10;
 
-      // loadObjFile.activateBloomEffect();
-      // loadObjFile.lightContainer[0].behavior.setOsc0(-2, 2, 0.001)
-      // loadObjFile.lightContainer[0].behavior.value_ = -1;
-      // loadObjFile.lightContainer[0].updater.push((light) => {
-      //   light.position[0] = light.behavior.setPath0()
-      //   light.target[0] = light.behavior.setPath0()
-      // })
+      loadObjFile.activateBloomEffect();
+      loadObjFile.lightContainer[0].behavior.setOsc0(-2, 2, 0.001)
+      loadObjFile.lightContainer[0].behavior.value_ = -1;
+      loadObjFile.lightContainer[0].updater.push((light) => {
+        light.position[0] = light.behavior.setPath0()
+        light.target[0] = light.behavior.setPath0()
+      })
 
       loadObjFile.lightContainer[0].position = [0, 17, -10];
       loadObjFile.lightContainer[0].target = [0, 0, -10];
@@ -153,7 +149,6 @@ export var loadObjFile = function() {
 
     loadObjFile.canvas.addEventListener("ray.hit.event", (e) => {
       console.log('ray.hit.event detected');
-      // if(e.detail.hitObject.morphTo) e.detail.hitObject.morphTo(0.0, 500);
     });
 
   })